% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportData.R
\docType{data}
\name{cwur.data}
\alias{cwur.data}
\title{ImportData}
\format{An object of class \code{data.frame} with 2200 rows and 14 columns.}
\usage{
cwur.data
}
\value{
2200 observations with 14 variables.
}
\description{
importing data from .csv file
}
\keyword{datasets}
