% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{GetData}
\alias{GetData}
\title{GetData}
\usage{
GetData()
}
\value{
table with 2200 obs. of 14 variables.
}
\description{
Data are already contained into the package UniversityRanking through the file cwur.RData.
}
