% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreInteractive.R, R/Score_Interactive.R
\name{ScoreInteractive}
\alias{ScoreInteractive}
\title{ScoreInteractive}
\usage{
ScoreInteractive(setpoint)

ScoreInteractive(setpoint)
}
\value{
plot

plot
}
\description{
This function takes a numeric argument to plot the scores of the universities over years.

This function takes a numeric argument to plot the scores of the universities over years.
}
