% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Top10_2015.R
\name{Top_nUniversities_year}
\alias{Top_nUniversities_year}
\title{Top_nUniversities_nYear}
\usage{
Top_nUniversities_year(nUni, nYear)
}
\value{
plot
}
\description{
The function Top_nUniversities_nYear plots the top desidered number of universities for a chosen year between 2012 and 2015.
}
