% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopUni.R
\name{TopUni}
\alias{TopUni}
\title{TopUni}
\usage{
TopUni(nUni, nYear)
}
\value{
plot
}
\description{
The function TopUni plots the top desidered number of universities for a chosen year between 2012 and 2015. It is possible to choose up to 1024 institutions.
}
\examples{
TopUni(10, 2016), TopUni(1200, 2013), TopUni(15, 2014)

}
