\name{cwur.data}
\alias{cwur.data}
\docType{data}
\title{ World University Rankings}
\description{This dataset contains the criterias which are ususally used to rank universites around the world.}
\usage{data("cwur.data")}
\format{A data frame with 2200 observations on the following 14 variables.
  \describe{\item{\code{world_rank}}{a numeric vector}
    \item{\code{institution}}{\code{Aalborg University}
    \code{Aalto University}
    \code{Aarhus University}
    \code{Aberystwyth University}
    \code{Åbo Akademi University}
    \code{Adam Mickiewicz University in Poznań}
    \code{AGH University of Science and Technology}
    \code{Ain Shams University}
    \code{Aix-Marseille University}
    \code{Ajou University}
    \code{Albany Medical College}
    \code{Albert Ludwig University of Freiburg}
    \code{Alexandria University}
    \code{All India Institute of Medical Sciences, New Delhi}
    \code{American University of Beirut}
    \code{Amirkabir University of Technology}
    \code{Ankara University}
    \code{Aristotle University of Thessaloniki}
    \code{Arizona State University}
    \code{Aston University}
    \code{Auburn University}
    \code{Australian National University}
    \code{Autonomous University of Barcelona}
    \code{Autonomous University of Madrid}
    \code{Babeș-Bolyai University}
    \code{Banaras Hindu University}
    \code{Bangor University}
    \code{Bar-Ilan University}
    \code{Baylor College of Medicine}
    \code{Baylor University}
    \code{Beihang University}
    \code{Beijing Institute of Technology}
    \code{Beijing Jiaotong University}
    \code{Beijing Normal University}
    \code{Beijing University of Chemical Technology}
    \code{Beijing University of Technology}
    \code{Ben-Gurion University of the Negev}
    \code{Bielefeld University}
    \code{Bilkent University}
    \code{Binghamton University}
    \code{Birkbeck, University of London}
    \code{Blaise Pascal University}
    \code{Boğaziçi University}
    \code{Boston College}
    \code{Boston University}
    \code{Bowling Green State University}
    \code{Brandeis University}
    \code{Braunschweig University of Technology}
    \code{Brigham Young University}
    \code{Brock University}
    \code{Brown University}
    \code{Brunel University} \code{Budapest University of Technology and Economics} \code{Cairo University} \code{California Institute of Technology} \code{Capital Medical University} \code{Cardiff University} \code{Carleton University} \code{Carlos III University of Madrid} \code{Carnegie Mellon University} \code{Case Western Reserve University} \code{Catholic University of Korea} \code{Catholic University of the Sacred Heart} \code{Central China Normal University} \code{Central South University} \code{Chalmers University of Technology} \code{Chang Gung University} \code{Charles Darwin University} \code{Charles University in Prague} \code{Chemnitz University of Technology} \code{Chiang Mai University} \code{Chiba University} \code{China Agricultural University} \code{China Medical University (PRC)} \code{China Medical University (Taiwan)} \code{China Pharmaceutical University} \code{China University of Geosciences (Wuhan)} \code{Chinese University of Hong Kong} \code{Chonbuk National University} \code{Chongqing University} \code{Chonnam National University} \code{Chosun University} \code{Chulalongkorn University} \code{Chung Shan Medical University} \code{Chung Yuan Christian University} \code{Chung-Ang University} \code{Chungbuk National University} \code{Chungnam National University} \code{City College of New York} \code{City University London} \code{City University of Hong Kong} \code{Clarkson University} \code{Claude Bernard University Lyon 1} \code{Clemson University} \code{College of William and Mary} \code{Colorado School of Mines} \code{Colorado State University - Fort Collins} \code{Columbia University} \code{Comenius University in Bratislava} \code{Complutense University of Madrid} \code{Concordia University} \code{Cornell University} \code{Cranfield University} \code{Creighton University} \code{Curtin University} \code{Czech Technical University in Prague} \code{Dalhousie University} \code{Dalian University of Technology} \code{Dankook University} \code{Darmstadt University of Technology} \code{Dartmouth College} \code{Deakin University} \code{Delft University of Technology} \code{Dokuz Eylül University} \code{Dongguk University} \code{Donghua University} \code{Dresden University of Technology} \code{Drexel University} \code{Dublin City University} \code{Duke University} \code{Durham University} \code{East Carolina University} \code{East China Normal University} \code{East China University of Science and Technology} \code{École centrale de Lyon} \code{École Centrale Paris} \code{École normale supérieure - Paris} \code{École normale supérieure de Cachan} \code{École normale supérieure de Lyon} \code{École Polytechnique} \code{École Polytechnique de Montréal} \code{Ege University} \code{Ehime University} \code{Eindhoven University of Technology} \code{Emory University} \code{Eötvös Loránd University} \code{Erasmus University Rotterdam} \code{Ewha Womans University} \code{Federal University of ABC} \code{Federal University of Bahia} \code{Federal University of Ceará} \code{Federal University of Minas Gerais} \code{Federal University of Paraná} \code{Federal University of Pernambuco} \code{Federal University of Rio de Janeiro} \code{Federal University of Rio Grande do Sul} \code{Federal University of Santa Catarina} \code{Federal University of Santa Maria} \code{Federal University of São Carlos} \code{Federal University of São Paulo} \code{Federico Santa María Technical University} \code{Feng Chia University} \code{Flinders University} \code{Florida Atlantic University} \code{Florida Institute of Technology} \code{Florida International University} \code{Florida State University} \code{Fluminense Federal University} \code{Fourth Military Medical University} \code{Free University of Berlin} \code{Fu Jen Catholic University} \code{Fudan University} \code{Fukuoka University} \code{Fuzhou University} \code{G. d'Annunzio University Chieti-Pescara} \code{Gachon University} \code{Gazi University} \code{George Mason University} \code{George Washington University} \code{Georgetown University} \code{Georgia Institute of Technology} \code{Georgia Regents University} \code{Georgia State University} \code{Ghent University} \code{Gifu University} \code{Goethe University Frankfurt} \code{Graduate University for Advanced Studies} \code{Graz University of Technology} \code{Grenoble Institute of Technology} \code{Griffith University} \code{Guangxi University} \code{Gunma University} \code{Gwangju Institute of Science and Technology} \code{Gyeongsang National University} \code{Hacettepe University} \code{Hallym University} \code{Hamamatsu University School of Medicine} \code{Hannover Medical School} \code{Hanyang University} \code{Harbin Engineering University} \code{Harbin Institute of Technology} \code{Harbin Medical University} \code{Harvard University} \code{Hasselt University} \code{Hebrew University of Jerusalem} \code{Hefei University of Technology} \code{Henan Normal University} \code{Heriot-Watt University} \code{Hirosaki University} \code{Hiroshima University} \code{Hokkaido University} \code{Hong Kong Baptist University} \code{Hong Kong Polytechnic University} \code{Hong Kong University of Science and Technology} \code{Howard University} \code{Huazhong Agricultural University} \code{Huazhong University of Science and Technology} \code{Humboldt University of Berlin} \code{Hunan Normal University} \code{Hunan University} \code{Hunter College} \code{I-Shou University} \code{Ibaraki University} \code{Icahn School of Medicine at Mount Sinai} \code{Illinois Institute of Technology} \code{Imperial College London} \code{Indian Institute of Science} \code{Indian Institute of Technology Bombay} \code{Indian Institute of Technology Delhi} \code{Indian Institute of Technology Kanpur} \code{Indian Institute of Technology Kharagpur} \code{Indian Institute of Technology Madras} \code{Indian Institute of Technology Roorkee} \code{Indiana University - Bloomington} \code{Indiana University-Purdue University Indianapolis} \code{Inha University} \code{Inje University} \code{Innsbruck Medical University} \code{Institut national des sciences Appliquées de Lyon} \code{International School for Advanced Studies} \code{Iowa State University} \code{Iran University of Science and Technology} \code{Isfahan University of Technology} \code{Istanbul Technical University} \code{Istanbul University} \code{Jacobs University Bremen} \code{Jadavpur University} \code{Jagiellonian University} \code{James Cook University} \code{Jaume I University} \code{Jawaharlal Nehru Centre for Advanced Scientific Research} \code{Jean Monnet University} \code{Jiangnan University} \code{Jiangsu University} \code{Jichi Medical University} \code{Jikei University School of Medicine} \code{Jilin University} \code{Jinan University} \code{Johannes Gutenberg University of Mainz} \code{Johannes Kepler University of Linz} \code{Johns Hopkins University} \code{Joseph Fourier University} \code{Julius-Maximilian's University of Würzburg} \code{Juntendo University} \code{Justus Liebig University Giessen} \code{Kagoshima University} \code{Kaiserslautern University of Technology} \code{Kanazawa University} \code{Kangwon National University} \code{Kansas State University} \code{Kaohsiung Medical University} \code{Karlsruhe Institute of Technology} \code{Karolinska Institute} \code{Katholieke Universiteit Leuven} \code{Keele University} \code{Keio University} \code{Kent State University} \code{Kindai University} \code{King Abdulaziz University} \code{King Abdullah University of Science and Technology} \code{King Fahd University of Petroleum and Minerals} \code{King Juan Carlos University} \code{King Saud University} \code{King's College London} \code{Kitasato University} \code{Kobe University} \code{Kōchi University} \code{Konkuk University} \code{Korea Advanced Institute of Science and Technology (KAIST)} \code{Korea University} \code{Kumamoto University} \code{Kurume University} \code{Kyoto Institute of Technology} \code{Kyoto Prefectural University of Medicine} \code{Kyoto University} \code{Kyung Hee University} \code{Kyungpook National University} \code{Kyushu University} \code{La Trobe University} \code{Lancaster University} \code{Lanzhou University} \code{Laval University} \code{Lehigh University} \code{Leiden University} \code{Leipzig University} \code{Lille 2 University of Health and Law} \code{Linköping University} \code{Liverpool John Moores University} \code{Loma Linda University} \code{Lomonosov Moscow State University} \code{London School of Economics} \code{London School of Hygiene & Tropical Medicine} \code{Loughborough University} \code{Louisiana State University - Baton Rouge} \code{Louisiana Tech University} \code{Loyola University Chicago} \code{LSU Health Sciences Center New Orleans} \code{Ludwig Maximilian University of Munich} \code{Lund University} \code{Maastricht University} \code{Macquarie University} \code{Magna Græcia University of Catanzaro} \code{Mahidol University} \code{Makerere University} \code{Manchester Metropolitan University} \code{Mansoura University} \code{Marche Polytechnic University} \code{Marquette University} \code{Martin Luther University of Halle-Wittenberg} \code{Masaryk University} \code{Massachusetts Institute of Technology} \code{Massey University} \code{McGill University} \code{McMaster University} \code{Medical College of Wisconsin} \code{Medical University of Graz} \code{Medical University of South Carolina} \code{Medical University of Vienna} \code{Memorial University of Newfoundland} \code{Miami University} \code{Michigan State University} \code{Michigan Technological University} \code{Middle East Technical University} \code{Mie University} \code{Miguel Hernández University of Elche} \code{Mines ParisTech} \code{Mississippi State University} \code{Missouri University of Science and Technology} \code{Monash University} \code{Montana State University - Bozeman} \code{Moscow Institute of Physics and Technology} \code{Murdoch University} \code{Nagasaki University} \code{Nagoya City University} \code{Nagoya Institute of Technology} \code{Nagoya University} \code{Nanjing Agricultural University} \code{Nanjing Medical University} \code{Nanjing Normal University} \code{Nanjing University} \code{Nanjing University of Aeronautics and Astronautics} \code{Nanjing University of Science and Technology} \code{Nanjing University of Technology} \code{Nankai University} \code{Nanyang Technological University} \code{Nara Institute of Science and Technology} \code{National and Kapodistrian University of Athens} \code{National Autonomous University of Mexico} \code{National Central University} \code{National Cheng Kung University} \code{National Chiao Tung University} \code{National Chung Cheng University} \code{National Chung Hsing University} \code{National Dong Hwa University} \code{National Polytechnic Institute} \code{National Polytechnic Institute of Toulouse} \code{National Research Nuclear University MEPhI} \code{National Sun Yat-sen University} \code{National Taipei University of Technology} \code{National Taiwan Normal University} \code{National Taiwan Ocean University} \code{National Taiwan University} \code{National Taiwan University of Science and Technology} \code{National Technical University of Athens} \code{National Tsing Hua University} \code{National University of Colombia} \code{National University of Córdoba} \code{National University of Defense Technology} \code{National University of Ireland, Galway} \code{National University of Ireland, Maynooth} \code{National University of La Plata} \code{National University of Rosario} \code{National University of Singapore} \code{National Yang-Ming University} \code{New Jersey Institute of Technology} \code{New Mexico State University} \code{New York Medical College} \code{New York University} \code{Newcastle University} \code{Nicolaus Copernicus University in Toruń} \code{Nihon University} \code{Niigata University} \code{North Carolina State University} \code{North Dakota State University - Main Campus} \code{Northeast Normal University} \code{Northeastern University} \code{Northeastern University (China)} \code{Northern Arizona University} \code{Northern Illinois University} \code{Northwest A&F University} \code{Northwest University (China)} \code{Northwestern Polytechnical University} \code{Northwestern University} \code{Norwegian University of Life Sciences} \code{Norwegian University of Science and Technology} \code{Novosibirsk State University} \code{Oakland University} \code{Ocean University of China} \code{Ohio State University, Columbus} \code{Ohio University} \code{Okayama University} \code{Oklahoma State University–Stillwater} \code{Old Dominion University} \code{Open University (UK)} \code{Örebro University} \code{Oregon Health & Science University} \code{Oregon State University} \code{Osaka City University} \code{Osaka Prefecture University} \code{Osaka University} \code{Otto-von-Guericke University Magdeburg} \code{Palacký University, Olomouc} \code{Panjab University} \code{Paris Descartes University} \code{Paris Diderot University - Paris 7} \code{Paris-Est Créteil Val-de-Marne University} \code{Paul Sabatier University - Toulouse III} \code{Paul Valéry University, Montpellier III} \code{Peking Union Medical College} \code{Peking University} \code{Pennsylvania State University, University Park} \code{Pierre-and-Marie-Curie University} \code{Plymouth University} \code{Pohang University of Science and Technology} \code{Polytechnic University of Catalonia} \code{Polytechnic University of Milan} \code{Polytechnic University of Turin} \code{Polytechnic University of Valencia} \code{Pompeu Fabra University} \code{Pontifical Catholic University of Chile} \code{Portland State University} \code{Princeton University} \code{Public University of Navarre} \code{Pukyong National University} \code{Purdue University, West Lafayette} \code{Pusan National University} \code{Putra University, Malaysia} \code{Queen Mary University of London} \code{Queen's University} \code{Queen's University Belfast} \code{Queens College, City University of New York} \code{Queensland University of Technology} \code{Radboud University Nijmegen} \code{Renmin University of China} \code{Rensselaer Polytechnic Institute} \code{Rice University} \code{Rio de Janeiro State University} \code{Ritsumeikan University} \code{RMIT University} \code{Rochester Institute of Technology} \code{Rockefeller University} \code{Roma Tre University} \code{Rovira i Virgili University} \code{Royal College of Surgeons in Ireland} \code{Royal Holloway, University of London} \code{Royal Institute of Technology} \code{Royal Veterinary College, University of London} \code{Ruhr University Bochum} \code{Ruprecht Karl University of Heidelberg} \code{Rush University} \code{Rutgers University-New Brunswick} \code{Rutgers, The State University of New Jersey - Newark} \code{RWTH Aachen University} \code{Ryerson University} \code{Saarland University} \code{Saga University} \code{Saint Louis University} \code{Saint Petersburg State University} \code{Saitama Medical University} \code{Saitama University} \code{San Diego State University} \code{San Francisco State University} \code{Sapienza University of Rome} \code{Sapporo Medical University} \code{Scuola Normale Superiore di Pisa} \code{Second Military Medical University} \code{Second University of Naples} \code{Sejong University} \code{Semmelweis University} \code{Seoul National University} \code{Shaanxi Normal University} \code{Shahid Beheshti University of Medical Sciences} \code{Shandong University} \code{Shanghai Jiao Tong University} \code{Shanghai Normal University} \code{Shanghai University} \code{Shanxi University} \code{Sharif University of Technology} \code{Shenzhen University} \code{Shimane University} \code{Shinshu University} \code{Shiraz University} \code{Shizuoka University} \code{Showa University} \code{Sichuan University} \code{Simon Fraser University} \code{Sofia University} \code{Sogang University} \code{Soochow University (Suzhou)} \code{South China Agricultural University} \code{South China Normal University} \code{South China University of Technology} \code{Southeast University} \code{Southern Illinois University Carbondale} \code{Southern Medical University} \code{Southern Methodist University} \code{Southwest Jiaotong University} \code{Southwest University} \code{St George's, University of London} \code{Stanford University} \code{State University of New York Upstate Medical University} \code{Stellenbosch University} \code{Stockholm University} \code{Stony Brook University} \code{Sun Yat-sen University} \code{Sungkyunkwan University} \code{SUNY Downstate Medical Center} \code{Swansea University} \code{Swedish University of Agricultural Sciences} \code{Swinburne University of Technology} \code{Swiss Federal Institute of Technology in Lausanne} \code{Swiss Federal Institute of Technology in Zurich} \code{Syracuse University} \code{Taipei Medical University} \code{Tamkang University} \code{Tampere University of Technology} \code{Tarbiat Modares University} \code{Tata Institute of Fundamental Research} \code{Technical University of Berlin} \code{Technical University of Denmark} \code{Technical University of Dortmund} \code{Technical University of Madrid} \code{Technical University of Munich} \code{Technion – Israel Institute of Technology} \code{Tehran University of Medical Sciences} \code{Teikyo University} \code{Tel Aviv University} \code{Temple University} \code{Texas A&M Health Science Center} \code{Texas A&M University, College Station} \code{Texas Tech University (TTU)} \code{The Catholic University of America} \code{Third Military Medical University} \code{Thomas Jefferson University} \code{Tianjin Medical University} \code{Tianjin University} \code{Tilburg University} \code{Toho University} \code{Tohoku University} \code{Tokai University} \code{Tokyo Institute of Technology} \code{Tokyo Medical and Dental University} \code{Tokyo Metropolitan University} \code{Tokyo University of Agriculture and Technology} \code{Tokyo University of Science} \code{Tokyo Women's Medical University} \code{Tongji University} \code{Tottori University} \code{Trent University} \code{Trinity College, Dublin} \code{Tsinghua University} \code{Tufts University} \code{Tulane University} \code{Ulsan National Institute of Science and Technology} \code{Umeå University} \code{UNESP, São Paulo State University} \code{United Arab Emirates University} \code{Universidade Nova de Lisboa} \code{Université catholique de Louvain} \code{Université de Namur} \code{Université libre de Bruxelles} \code{University at Albany, SUNY} \code{University at Buffalo, The State University of New York} \code{University College Cork} \code{University College Dublin} \code{University College London} \code{University François Rabelais, Tours} \code{University Lille 1: Sciences and Technologies} \code{University of A Coruña} \code{University of Aberdeen} \code{University of Adelaide} \code{University of Akron} \code{University of Alabama - Tuscaloosa} \code{University of Alabama at Birmingham} \code{University of Alabama in Huntsville} \code{University of Alaska Fairbanks} \code{University of Alberta} \code{University of Alcalá} \code{University of Alicante} \code{University of Almería} \code{University of Amsterdam} \code{University of Angers} \code{University of Antwerp} \code{University of Arizona} \code{University of Arkansas - Fayetteville} \code{University of Arkansas for Medical Sciences} \code{University of Auckland} \code{University of Augsburg} \code{University of Aveiro} \code{University of Barcelona} \code{University of Bari Aldo Moro} \code{University of Basel} \code{University of Basilicata} \code{University of Bath} \code{University of Bayreuth} \code{University of Belgrade} \code{University of Bergen} \code{University of Bern} \code{University of Birmingham} \code{University of Bologna} \code{University of Bonn} \code{University of Bordeaux} \code{University of Bordeaux I} \code{University of Bordeaux II} \code{University of Bradford} \code{University of Brasília} \code{University of Bremen} \code{University of Brescia} \code{University of Brighton} \code{University of Bristol} \code{University of British Columbia} \code{University of Bucharest} \code{University of Buenos Aires} \code{University of Burgundy} \code{University of Cádiz} \code{University of Caen Lower Normandy} \code{University of Cagliari} \code{University of Calabria} \code{University of Calcutta} \code{University of Calgary} \code{University of California, Berkeley} \code{University of California, Davis} \code{University of California, Irvine} \code{University of California, Los Angeles} \code{University of California, Merced} \code{University of California, Riverside} \code{University of California, San Diego} \code{University of California, San Francisco} \code{University of California, Santa Barbara} \code{University of California, Santa Cruz} \code{University of Cambridge} \code{University of Camerino} \code{University of Campinas} \code{University of Cantabria} \code{University of Canterbury} \code{University of Cape Town} \code{University of Castilla–La Mancha} \code{University of Catania} \code{University of Central Florida} \code{University of Central Lancashire} \code{University of Chemistry and Technology, Prague} \code{University of Chicago} \code{University of Chile} \code{University of Cincinnati} \code{University of Coimbra} \code{University of Cologne} \code{University of Colorado Boulder} \code{University of Colorado Denver} \code{University of Concepción} \code{University of Connecticut} \code{University of Copenhagen} \code{University of Córdoba (Spain)} \code{University of Crete} \code{University of Cyprus} \code{University of Dayton} \code{University of Debrecen} \code{University of Delaware} \code{University of Delhi} \code{University of Denver} \code{University of Duisburg-Essen} \code{University of Dundee} \code{University of Düsseldorf} \code{University of East Anglia} \code{University of Eastern Finland} \code{University of Eastern Piedmont Amedeo Avogadro} \code{University of Edinburgh} \code{University of Electro-Communications} \code{University of Electronic Science and Technology of China} \code{University of Erlangen-Nuremberg} \code{University of Essex} \code{University of Évry Val d'Essonne} \code{University of Exeter} \code{University of Extremadura} \code{University of Ferrara} \code{University of Florence} \code{University of Florida} \code{University of Franche-Comté} \code{University of Fribourg} \code{University of Fukui} \code{University of Geneva} \code{University of Genoa} \code{University of Georgia} \code{University of Girona} \code{University of Glasgow} \code{University of Gothenburg} \code{University of Göttingen} \code{University of Granada} \code{University of Graz} \code{University of Greifswald} \code{University of Groningen} \code{University of Guelph} \code{University of Haifa} \code{University of Hamburg} \code{University of Hanover} \code{University of Hawaii at Manoa} \code{University of Helsinki} \code{University of Hertfordshire} \code{University of Hohenheim} \code{University of Hong Kong} \code{University of Houston} \code{University of Hull} \code{University of Hyderabad} \code{University of Hyogo} \code{University of Iceland} \code{University of Idaho} \code{University of Illinois at Chicago} \code{University of Illinois at Urbana–Champaign} \code{University of Innsbruck} \code{University of Insubria} \code{University of Ioannina} \code{University of Iowa} \code{University of Jaén} \code{University of Jena} \code{University of Jyväskylä} \code{University of Kansas} \code{University of Kent} \code{University of Kentucky} \code{University of Kiel} \code{University of Konstanz} \code{University of KwaZulu-Natal} \code{University of L'Aquila} \code{University of La Laguna} \code{University of Lausanne} \code{University of Leeds} \code{University of Leicester} \code{University of Lethbridge} \code{University of Liège} \code{University of Limerick} \code{University of Lisbon} \code{University of Liverpool} \code{University of Ljubljana} \code{University of Lleida} \code{University of Lorraine} \code{University of Los Andes (Colombia)} \code{University of Louisville} \code{University of Lübeck} \code{University of Maine (Le Mans)} \code{University of Maine, Orono} \code{University of Málaga} \code{University of Malaya} \code{University of Manchester} \code{University of Manitoba} \code{University of Mannheim} \code{University of Marburg} \code{University of Maribor} \code{University of Maryland, Baltimore} \code{University of Maryland, Baltimore County} \code{University of Maryland, College Park} \code{University of Massachusetts Amherst} \code{University of Massachusetts Boston} \code{University of Massachusetts Lowell} \code{University of Massachusetts Medical School} \code{University of Melbourne} \code{University of Memphis} \code{University of Messina} \code{University of Miami} \code{University of Michigan, Ann Arbor} \code{University of Milan} \code{University of Milan - Bicocca} \code{University of Minho} \code{University of Minnesota, Twin Cities} \code{University of Mississippi - Oxford Campus} \code{University of Mississippi Medical Center} \code{University of Missouri–Columbia} \code{University of Missouri–Kansas City} \code{University of Missouri–St. Louis} \code{University of Miyazaki} \code{University of Modena and Reggio Emilia} \code{University of Mons} \code{University of Montana - Missoula} \code{University of Montpellier} \code{University of Montpellier 1} \code{University of Montpellier 2} \code{University of Montreal} \code{University of Münster} \code{University of Murcia} \code{University of Nantes} \code{University of Naples Federico II} \code{University of Natural Resources and Life Sciences, Vienna} \code{University of Navarra} \code{University of Nebraska Medical Center} \code{University of Nebraska Omaha} \code{University of Nebraska–Lincoln} \code{University of Neuchâtel} \code{University of Nevada, Las Vegas} \code{University of Nevada, Reno} \code{University of New Brunswick} \code{University of New England (Australia)} \code{University of New Hampshire} \code{University of New Mexico} \code{University of New Orleans} \code{University of New South Wales} \code{University of Newcastle} \code{University of Nice Sophia Antipolis} \code{University of North Carolina at Chapel Hill} \code{University of North Carolina at Charlotte} \code{University of North Carolina at Greensboro} \code{University of North Dakota} \code{University of North Texas, Denton} \code{University of Notre Dame} \code{University of Nottingham} \code{University of Oklahoma - Norman Campus} \code{University of Oklahoma Health Sciences Center} \code{University of Oldenburg} \code{University of Oregon} \code{University of Orléans} \code{University of Oslo} \code{University of Osnabrück} \code{University of Otago} \code{University of Ottawa} \code{University of Oulu} \code{University of Oviedo} \code{University of Oxford} \code{University of Padua} \code{University of Palermo} \code{University of Paris 13} \code{University of Paris-Sud} \code{University of Parma} \code{University of Patras} \code{University of Pau and Pays de l'Adour} \code{University of Pavia} \code{University of Pécs} \code{University of Pennsylvania} \code{University of Perpignan Via Domitia} \code{University of Perugia} \code{University of Picardie Jules Verne} \code{University of Pisa} \code{University of Pittsburgh - Pittsburgh Campus} \code{University of Poitiers} \code{University of Porto} \code{University of Portsmouth} \code{University of Potsdam} \code{University of Pretoria} \code{University of Puerto Rico at Mayagüez} \code{University of Québec at Montreal} \code{University of Queensland} \code{University of Reading} \code{University of Regensburg} \code{University of Regina} \code{University of Reims Champagne-Ardenne} \code{University of Rennes 1} \code{University of Rhode Island} \code{University of Rochester} \code{University of Rome Tor Vergata} \code{University of Rostock} \code{University of Rouen} \code{University of Salamanca} \code{University of Salento} \code{University of Salerno} \code{University of Salzburg} \code{University of Santiago de Compostela} \code{University of São Paulo} \code{University of Saskatchewan} \code{University of Sassari} \code{University of Savoy} \code{University of Science and Technology Beijing} \code{University of Science and Technology of China} \code{University of Science, Malaysia} \code{University of Seoul} \code{University of Seville} \code{University of Sheffield} \code{University of Sherbrooke} \code{University of Siegen} \code{University of Siena} \code{University of Silesia} \code{University of South Alabama} \code{University of South Australia} \code{University of South Carolina - Columbia} \code{University of South Florida - Tampa} \code{University of Southampton} \code{University of Southern California} \code{University of Southern Denmark} \code{University of Southern Mississippi} \code{University of St Andrews} \code{University of Stirling} \code{University of Strasbourg} \code{University of Strathclyde} \code{University of Stuttgart} \code{University of Surrey} \code{University of Sussex} \code{University of Sydney} \code{University of Szeged} \code{University of Tampere} \code{University of Tartu} \code{University of Tasmania} \code{University of Technology, Sydney} \code{University of Tehran} \code{University of Tennessee, Knoxville} \code{University of Texas at Arlington} \code{University of Texas at Austin} \code{University of Texas at Dallas} \code{University of Texas at El Paso} \code{University of Texas at San Antonio} \code{University of Texas Health Science Center at Houston} \code{University of Texas Health Science Center at San Antonio} \code{University of Texas MD Anderson Cancer Center} \code{University of Texas Medical Branch at Galveston} \code{University of Texas Southwestern Medical Center} \code{University of the Algarve} \code{University of the Balearic Islands} \code{University of the Basque Country} \code{University of the Republic (Uruguay)} \code{University of the Ryukyus} \code{University of the Witwatersrand} \code{University of Thessaly} \code{University of Tokushima} \code{University of Tokyo} \code{University of Toledo} \code{University of Toronto} \code{University of Toyama} \code{University of Trento} \code{University of Trieste} \code{University of Tromsø - The Arctic University of Norway} \code{University of Tsukuba} \code{University of Tübingen} \code{University of Turin} \code{University of Turku} \code{University of Twente} \code{University of Udine} \code{University of Ulm} \code{University of Ulsan} \code{University of Ulster} \code{University of Utah} \code{University of Valencia} \code{University of Valenciennes and Hainaut-Cambresis} \code{University of Valladolid} \code{University of Vermont} \code{University of Verona} \code{University of Veterinary Medicine Vienna} \code{University of Victoria} \code{University of Vienna} \code{University of Vigo} \code{University of Virginia} \code{University of Waikato} \code{University of Wales, Trinity Saint David} \code{University of Warsaw} \code{University of Warwick} \code{University of Washington - Seattle} \code{University of Waterloo} \code{University of Western Australia} \code{University of Western Brittany} \code{University of Western Sydney} \code{University of Windsor} \code{University of Wisconsin–Madison} \code{University of Wisconsin–Milwaukee} \code{University of Wollongong} \code{University of Wrocław} \code{University of Wyoming} \code{University of Yamanashi} \code{University of York} \code{University of Zagreb} \code{University of Zaragoza} \code{University of Zurich} \code{Uppsala University} \code{Utah State University} \code{Utrecht University} \code{Vanderbilt University} \code{Versailles Saint-Quentin-en-Yvelines University} \code{Victoria University of Wellington} \code{Vienna University of Technology} \code{Vilnius University} \code{Virginia Commonwealth University} \code{Virginia Polytechnic Institute and State University} \code{Vita-Salute San Raffaele University} \code{Vrije Universiteit Brussel} \code{VU University Amsterdam} \code{Wageningen University and Research Centre} \code{Wake Forest University} \code{Warsaw University of Technology} \code{Waseda University} \code{Washington State University, Pullman} \code{Washington University in St. Louis} \code{Wayne State University} \code{Weizmann Institute of Science} \code{Wesleyan University} \code{West Virginia University} \code{Western Michigan University} \code{Western University (The University of Western Ontario)} \code{Wilfrid Laurier University} \code{Williams College} \code{Wright State University - Dayton} \code{Wrocław University of Technology} \code{Wuhan University} \code{Wuhan University of Technology} \code{Xian Jiaotong University} \code{Xiamen University} \code{Xiangtan University} \code{Xidian University} \code{Yale University} \code{Yamagata University} \code{Yamaguchi University} \code{Yanbian University} \code{Yangzhou University} \code{Yeshiva University} \code{Yeungnam University} \code{Yokohama City University} \code{Yokohama National University} \code{Yonsei University} \code{York University} \code{Yuan Ze University} \code{Zhejiang Normal University} \code{Zhejiang University} \code{Zhejiang University of Technology} \code{Zhengzhou University}}
    \item{\code{country}}{
    \code{Argentina},
    \code{Australia},
    \code{Austria},
    \code{Belgium}.
    \code{Brazil}, \code{Bulgaria}, \code{Canada}, \code{Chile} \code{China} \code{Colombia} \code{Croatia} \code{Cyprus} \code{Czech Republic} \code{Denmark} \code{Egypt} \code{Estonia} \code{Finland} \code{France} \code{Germany} \code{Greece} \code{Hong Kong} \code{Hungary} \code{Iceland} \code{India} \code{Iran} \code{Ireland} \code{Israel} \code{Italy} \code{Japan} \code{Lebanon} \code{Lithuania} \code{Malaysia} \code{Mexico} \code{Netherlands} \code{New Zealand} \code{Norway} \code{Poland} \code{Portugal} \code{Puerto Rico} \code{Romania} \code{Russia} \code{Saudi Arabia} \code{Serbia} \code{Singapore} \code{Slovak Republic} \code{Slovenia} \code{South Africa} \code{South Korea} \code{Spain} \code{Sweden} \code{Switzerland} \code{Taiwan} \code{Thailand} \code{Turkey} \code{Uganda} \code{United Arab Emirates} \code{United Kingdom} \code{Uruguay} \code{USA}}

    \item{\code{national_rank}}{a numeric vector}
    \item{\code{quality_of_education}}{a numeric vector}
    \item{\code{alumni_employment}}{a numeric vector}
    \item{\code{quality_of_faculty}}{a numeric vector}
    \item{\code{publications}}{a numeric vector}
    \item{\code{influence}}{a numeric vector}
    \item{\code{citations}}{a numeric vector}
    \item{\code{broad_impact}}{a numeric vector}
    \item{\code{patents}}{a numeric vector}
    \item{\code{score}}{a numeric vector}
    \item{\code{year}}{a numeric vector}
  }
}
\source{https://www.kaggle.com/mylesoneill/world-university-rankings}
\references{https://cwur.org}
\examples{
data(cwur.data)
str(cwur.data)
plot(cwur.data)}
\keyword{datasets}
